correcao em produto, atuailizacao da facade e finalizacao US3_2

+++ myFood-P2-2026/src/br/ufal/ic/p2/myfood/Modelos/Produtos/Produto.java
@@ -43,7 +43,7 @@ public class Produto {
         this.nome = nome;
     }
 
-    public double getValor() {
+    public float getValor() {
         return valor;
     }
